Re-baseline test, diff introduced by this PR is expected
The old test was using linear elasticity, but turned the
NonlinearGeometry flag ON, which is a nonsensical state. I have
now shifted all integrals to always occur in the reference
configuration (hence using Piola stress), so the solution to this
problem is expected to change.

On a separate note, this test could be better. It's checking the
l2 norm of the nodal solution vector against a simulation-obersved
baseline value. So the test is only clocking diffs from a specific
commit, without knowing if that behavior is correct.