Fix bug in domains created by Entire{Domain,Boundary} functions
The bug was that the mfem entity id vectors (eg, mfem_edge_ids_)
were not populated, which means that the dof_list function was
unable to find any dofs. Even worse, no error was thrown, you
would just get an empty dof_list. The calculation would continue,
but with the dofs left out of whatever operations you used.

This commit fixes the bug by removing the repeated code in the
Entire* functions. Now they use the factory functions used
for domain creation with spatial predicates, and just send a
predicate that is always true. The only downside is that the
Entire* functions now need a template paramater for the spatial
dimension. Seems a good trade-off to me.